ONEPLUS-150
Removed the functionality to post and view discussion comments when answering a question.
Added pad-top style class to div containing about, terms, etc in main.html. Because the submit button sits on top of it in preview page
Corrected the variable name that prevented discsussion comments from being displayed in next, right and wrong pages.